LayoutTests/fast/encoding/legacy-ios-encodings-expected.txt

This test verifies that certain legacy iOS text encodings are not supported.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


The following encodings should not be supported
PASS new TextDecoder("macos-6-10.2").encoding threw exception RangeError: Bad value.
PASS new TextDecoder("macos-6_2-10.4").encoding threw exception RangeError: Bad value.
PASS new TextDecoder("macos-35-10.2").encoding threw exception RangeError: Bad value.
PASS new TextDecoder("macos-29-10.2").encoding threw exception RangeError: Bad value.
PASS new TextDecoder("macos-7_3-10.2").encoding threw exception RangeError: Bad value.
PASS new TextDecoder("softbank-sjis").encoding threw exception RangeError: Bad value.
PASS new TextDecoder("x-mac-greek").encoding threw exception RangeError: Bad value.
PASS new TextDecoder("x-mac-turkish").encoding threw exception RangeError: Bad value.
PASS new TextDecoder("x-mac-centraleurroman").encoding threw exception RangeError: Bad value.
PASS new TextDecoder("windows-10006").encoding threw exception RangeError: Bad value.
PASS new TextDecoder("macgr").encoding threw exception RangeError: Bad value.
PASS new TextDecoder("x-macgreek").encoding threw exception RangeError: Bad value.
PASS new TextDecoder("windows-10081").encoding threw exception RangeError: Bad value.
PASS new TextDecoder("mactr").encoding threw exception RangeError: Bad value.
PASS new TextDecoder("x-macturkish").encoding threw exception RangeError: Bad value.
PASS new TextDecoder("windows-10029").encoding threw exception RangeError: Bad value.
PASS new TextDecoder("x-mac-ce").encoding threw exception RangeError: Bad value.
PASS new TextDecoder("macce").encoding threw exception RangeError: Bad value.
PASS new TextDecoder("maccentraleurope").encoding threw exception RangeError: Bad value.
PASS new TextDecoder("x-maccentraleurope").encoding threw exception RangeError: Bad value.
The following encoding names are supported by WebKit cross-platform, but some may be removed at a later time.
PASS new TextDecoder("x-mac-cyrillic").encoding is "x-mac-cyrillic"
PASS new TextDecoder("windows-10007").encoding is "x-mac-cyrillic"
PASS new TextDecoder("mac-cyrillic").encoding is "x-mac-cyrillic"
PASS new TextDecoder("maccy").encoding is "x-mac-cyrillic"
PASS new TextDecoder("x-maccyrillic").encoding is "x-mac-cyrillic"
PASS new TextDecoder("x-macukraine").encoding is "x-mac-cyrillic"
PASS successfullyParsed is true

TEST COMPLETE
